workingOnItServerAi/requirements.txt
Flask
flask-cors
python-dotenv
requests
google-generativeai
gunicorn

